fix(core): 修复运行时桥接与 HTTP 服务器稳定性
Runtime Bridge:
- \n 定界帧解析替代全缓冲 JSON 解析，避免粘包跨帧解析失败
- next_id_ 改为 int64_t，消除 32 位回绕风险
- <chrono> 精准超时替代手工 elapsed 累加
- err 变量命名冲突 → err_dict 修复
- RuntimeBridge::read_response 内联 try_parse_line lambda
- game_bridge: if-else 键查找链替换为 constexpr 查找表
- game_bridge: BUFFER_LIMIT 改为 static constexpr
- editor_plugin: 添加 is_bridge_connected() 访问器

HTTP Server:
- PollGuard 改为 RAII 构造即设置，消除遗漏复位路径
- subarray → resize + memcpy 修复 PackedByteArray subarray 问题
- timeout_msec_ 成员变量 → kTimeoutMsec 编译时常量
- http_parser: parse_utf8 错误码检查，read_buf.clear() 防止累积
- http_sse: SSE comment 尾缀 \r\n\r\n → \r\n（SSE 规范修复）
- mcp_handler: 版本表改为 constexpr std::array
  random_device 异常保护（某些环境如容器中会抛异常）
  operator bool() 添加 get_type() 守卫避免 Variant 空值 UB

+++ extensions/src/runtime/bridge.hpp
@@ -36,7 +36,7 @@ private:
     Status status_ = DISCONNECTED;
     godot::Ref<godot::StreamPeerTCP> tcp_;
     int port_ = 9601;
-    int next_id_ = 1;
+    int64_t next_id_ = 1;
     unsigned long long connecting_since_ = 0;
 };
